# Optimize release notes generation to use GitHub Milestones