--- a/go/tools/release-notes/release_notes.go
+++ b/go/tools/release-notes/release_notes.go
@@ -81,7 +81,6 @@ type (
 
 var (
 	releaseNotesPath = `changelog/`
-	numberOfThreads  = 10
 )
 
 const (
@@ -338,7 +337,6 @@ func main() {
 	)
 	pflag.StringVarP(&versionName, "version", "v", "", "name of the version (has to be the following format: v11.0.0)")
 	pflag.StringVarP(&summaryFile, "summary", "s", "", "readme file on which there is a summary of the release")
-	pflag.IntVar(&numberOfThreads, "threads", numberOfThreads, "Define the number of threads used to fetch data from GitHub's API. Lower this number if you hit request limit errors.")
 	pflag.Parse()
 
 	// The -version flag must be of a valid format.