## Description

This PR changes how we generate releases notes' changelog. We used to parse the `git log` output, this led to several errors in the past releases where some PRs made on a totally unrelated branch appeared in the changelog of a release.

With this PR we now use the GitHub Milestones to fetch the full list of merged Pull Request for a given release. This change significantly improves the time taken by the tool to generate the release notes, below is a before / after when generating the `v17.0.0` release notes:

```shell
# before
$> time go run ./go/tools/release-notes/release_notes.go -f v16.0.2 -t v17.0.0 --version v17.0.0
Found 460 merged PRs. Loading PR info................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................
go run ./go/tools/release-notes/release_notes.go -f v16.0.2 -t v17.0.0    31.70s user 20.03s system 184% cpu 28.061 total
```

```shell
# after
$> time go run ./go/tools/release-notes/release_notes.go --version v17.0.0 
go run ./go/tools/release-notes/release_notes.go --version v17.0.0  0.31s user 0.23s system 13% cpu 4.004 total
```

Going from 28 seconds to 4 seconds, there is a **~85% time improvement** to run the script. Moreover this new approach will allow us to never reach API request limit as **only one API call is made per execution,** instead of previously two API calls per Pull Requests we had in a release, potentially 460 PRs * 2 API call = **920 API calls**.

## Checklist

-   [x] "Backport to:" labels have been added if this change should be back-ported
-   [x] Tests were added or are not required
-   [x] Did the new or modified tests pass consistently locally and on the CI
-   [x] Documentation was added or is not required

## Deployment Notes

<!-- Notes regarding deployment of the contained body of work. These should note any db migrations, etc. -->